renamed to SimpleRelationship
Avoids a breaking change
