+++ packages/tc-ui/src/pages/edit/browser.js
@@ -1,7 +1,7 @@
 require('./main.css');
 const { init } = require('@financial-times/tc-schema-sdk');
 const {
-	RichRelationship: { withEditComponent: attachRichRelationshipPicker },
+	Relationship: { withEditComponent: attachRelationshipPicker },
 	LargeText: { withEditComponent: attachDocumentEditor },
 } = require('../../primitives/browser');
 
@@ -21,11 +21,9 @@ const initDocumentEditors = () => {
 const initRelationshipSelectors = entireRecord => {
 	[
 		...document.querySelectorAll(
-			'[data-component="rich-relationship-picker"]:not([data-disabled])',
+			'[data-component="relationship"]:not([data-disabled])',
 		),
-	].forEach(container =>
-		attachRichRelationshipPicker(container, entireRecord),
-	);
+	].forEach(container => attachRelationshipPicker(container, entireRecord));
 };
 
 // TODO - the relationship editors should expose their bad state in some way
@@ -37,7 +35,7 @@ const preventBadSubmission = () => {
 		.addEventListener('submit', ev => {
 			const editorsInUnselectedState = [
 				...document.querySelectorAll(
-					'[data-component="rich-relationship-picker"][data-is-unresolved]',
+					'[data-component="relationship"][data-is-unresolved]',
 				),
 			];
 			if (editorsInUnselectedState.length) {
