+++ packages/tc-ui/src/primitives/browser.js
@@ -1,9 +1,9 @@
 const Relationship = require('./relationship/browser');
-const RichRelationship = require('./rich-relationship/browser');
+const SimpleRelationship = require('./simple-relationship/browser');
 const LargeText = require('./large-text/browser');
 
 module.exports = {
 	Relationship,
-	RichRelationship,
+	SimpleRelationship,
 	LargeText,
 };
