+++ packages/tc-ui/src/primitives/relationship/browser.jsx
@@ -1,16 +1,18 @@
 require('../relationship-picker/main.css');
+require('./main.css');
 const React = require('react');
 const { render } = require('react-dom');
 const { RelationshipPicker } = require('../relationship-picker');
 const { SelectedRelationship } = require('./lib/selected-relationship');
 
 module.exports = {
-	withEditComponent: container =>
+	withEditComponent: (container, entireRecord) =>
 		render(
 			<RelationshipPicker
 				{...JSON.parse(container.dataset.props)}
+				entireRecord={entireRecord}
 				SelectedRelationship={SelectedRelationship}
-				componentName="relationship-picker"
+				componentName="relationship"
 			/>,
 			container.parentNode,
 		),
