+++ packages/tc-ui/src/primitives/relationship/server.jsx
@@ -1,5 +1,6 @@
 const React = require('react');
 const { getType } = require('@financial-times/tc-schema-sdk');
+const sortBy = require('lodash.sortby');
 const { WrappedEditComponent } = require('../../lib/components/input-wrapper');
 const { RelationshipPicker } = require('../relationship-picker');
 const { SelectedRelationship } = require('./lib/selected-relationship');
@@ -9,7 +10,7 @@ const RelationshipPickerContainer = props => (
 		<RelationshipPicker
 			{...props}
 			SelectedRelationship={SelectedRelationship}
-			componentName="relationship-picker"
+			componentName="relationship"
 		/>
 	</div>
 );
@@ -18,14 +19,6 @@ const {
 	ViewRelationship,
 	setRelationshipAnnotator,
 } = require('./lib/view-relationship');
-
-const maybeSort = (hasMany, props) => {
-	if (!hasMany) {
-		return '';
-	}
-	const sortField = props.has('name') ? 'name' : 'code';
-	return `(orderBy: ${sortField}_asc)`;
-};
 
 module.exports = {
 	name: 'Relationship',
@@ -37,23 +30,46 @@ module.exports = {
 			{...props}
 		/>
 	),
-	parser: value => {
-		if (!value) {
+	parser: (relValues, relProperties, assignComponent) => {
+		if (!relValues) {
 			return null;
 		}
-		value = JSON.parse(value);
+		relValues = JSON.parse(relValues);
 		// TODO use hasValue
-		if (!value) {
+		if (!relValues) {
 			return null;
 		}
-		return Array.isArray(value)
-			? value.map(({ code }) => code)
-			: value.code;
+		const isArray = Array.isArray(relValues);
+		const parsedRelValues = isArray
+			? relValues.map(({ code }) => ({ code }))
+			: { code: relValues.code };
+
+		if (assignComponent && relProperties) {
+			Object.entries(relProperties).forEach(([fieldName, fieldProps]) => {
+				const { parser } = assignComponent(fieldProps);
+				if (isArray) {
+					relValues.forEach((value, index) =>
+						Object.assign(parsedRelValues[index], {
+							[fieldName]: value[fieldName]
+								? parser(value[fieldName])
+								: null,
+						}),
+					);
+				} else {
+					Object.assign(parsedRelValues, {
+						[fieldName]: relValues[fieldName]
+							? parser(relValues[fieldName])
+							: null,
+					});
+				}
+			});
+		}
+		return parsedRelValues;
 	},
 	hasValue: (value, { hasMany }) =>
 		hasMany ? value && value.length : !!value,
 	setRelationshipAnnotator,
-	graphqlFragment: (propName, { type, hasMany }) => {
+	graphqlFragment: (propName, { type, properties }) => {
 		const typeDef = getType(type);
 		const props = new Set(['code']);
 		if (typeDef.properties.name) {
@@ -65,11 +81,43 @@ module.exports = {
 
 		Object.entries(typeDef.properties)
 			.filter(([, { useInSummary }]) => useInSummary)
-			.forEach(([name]) => props.add(name));
-
-		return `${propName} ${maybeSort(hasMany, props)} {${[...props].join(
+			.forEach(([name, { type: fieldType }]) =>
+				props.add(
+					['DateTime', 'Date', 'Time'].includes(fieldType)
+						? `${name} { formatted }`
+						: name,
+				),
+			);
+		const nodeProps = [...props].join(' ');
+		const relationshipProps = [...new Set(Object.keys(properties))].join(
 			' ',
-		)}}`;
+		);
+		return `${propName}_rel {${type} {${nodeProps}} ${relationshipProps}}`;
 	},
-	prepareValueForEdit: value => value || null,
+	prepareValueForEdit: (value, propDef) => {
+		if (propDef.hasMany) {
+			return value
+				? sortBy(value, `${propDef.type}.code`).map(item => ({
+						code: item.code || item[propDef.type].code,
+						name:
+							item.name ||
+							item.code ||
+							item[propDef.type].name ||
+							item[propDef.type].code,
+						...item,
+				  }))
+				: [];
+		}
+		return value
+			? {
+					code: value.code || value[propDef.type].code,
+					name:
+						value.name ||
+						value.code ||
+						value[propDef.type].name ||
+						value[propDef.type].code,
+					...value,
+			  }
+			: null;
+	},
 };
