+++ packages/tc-ui/src/primitives/simple-relationship/browser.jsx
@@ -1,18 +1,16 @@
 require('../relationship-picker/main.css');
-require('./main.css');
 const React = require('react');
 const { render } = require('react-dom');
 const { RelationshipPicker } = require('../relationship-picker');
 const { SelectedRelationship } = require('./lib/selected-relationship');
 
 module.exports = {
-	withEditComponent: (container, entireRecord) =>
+	withEditComponent: container =>
 		render(
 			<RelationshipPicker
 				{...JSON.parse(container.dataset.props)}
-				entireRecord={entireRecord}
 				SelectedRelationship={SelectedRelationship}
-				componentName="rich-relationship-picker"
+				componentName="simple-relationship"
 			/>,
 			container.parentNode,
 		),
